Updated Flutter version and reorganized library files according to Dart guidelines
