+++ src/pubspec.yaml
@@ -16,3 +16,6 @@ dev_dependencies:
 dependencies:
   flutter:
     sdk: flutter
+
+environment:
+  sdk: ">=2.0.0-dev.28.0 <3.0.0"
